+++ src/main/java/software/mari/flow_server/repository/PlatformRepository.java
@@ -1,7 +1,7 @@
 package software.mari.flow_server.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import software.mari.flow_server.model.product.access.Platform;
+import software.mari.flow_server.model.product.Platform;
 
-public interface PlatformRepository extends JpaRepository<Platform, Integer> {
+public interface PlatformRepository extends JpaRepository<Platform, Long> {
 }
