+++ src/main/java/software/mari/flow_server/repository/ProductCategoryRepository.java
@@ -3,5 +3,5 @@ package software.mari.flow_server.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import software.mari.flow_server.model.product.ProductCategory;
 
-public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Integer> {
+public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Long> {
 }
