+++ src/main/java/software/mari/flow_server/repository/ProductHighlightRepository.java
@@ -1,7 +1,7 @@
 package software.mari.flow_server.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import software.mari.flow_server.model.product.highlight.ProductHighlight;
+import software.mari.flow_server.model.product.ProductHighlight;
 
-public interface ProductHighlightRepository extends JpaRepository<ProductHighlight, Integer> {
+public interface ProductHighlightRepository extends JpaRepository<ProductHighlight, Long> {
 }
